\name{mime_type}
\alias{mime_type}
\title{Compute mime type from file extension.
Needs mechanism for user extension.}
\usage{
  mime_type(path)
}
\arguments{
  \item{path}{path to file}
}
\value{
  mime string
}
\description{
  Compute mime type from file extension. Needs mechanism
  for user extension.
}

