\name{render}
\alias{render}
\title{Send output to browser...}
\usage{render(text, status=200, mime_type="text/html", ...)}

\description{
  Send output to browser
}
\arguments{
  \item{text}{a character vector}
  \item{status}{http status code}
  \item{mime_type}{mime type}
  \item{...}{other named arguments become other http headers}
}
