\name{route_re}
\alias{route_re}
\title{Parse route (with parameters) into regular expression.}
\usage{
  route_re(route)
}
\arguments{
  \item{route}{route url.}
}
\value{
  \item{match}{regular expression to match route}
  \item{params}{parameter names}
}
\description{
  Parse route (with parameters) into regular expression.
}

