\name{render}
\alias{render}
\title{Send output to browser...}
\usage{render(text, status=200, mime_type="text/html", headers=c())}

\description{
  Send output to browser
}
\arguments{
  \item{text}{a character vector}
  \item{status}{http status code}
  \item{mime_type}{mime type}
  \item{headers}{character vector of http headers}
}
