\name{render_brew}
\alias{render_brew}
\title{Render template using brew.}
\usage{render_brew(template, params=list(), path=getwd())}

\description{
  Render template using brew. Templates are located in \code{views}.
}
\arguments{
  \item{template}{string giving template name (file name without extension)}
  \item{params}{list of parameters to be evaluated in template}
  \item{path}{web app path}
}
