\name{route_re}
\alias{route_re}
\title{Parse route (with parameters) into regular expression.}
\usage{route_re(route)}

\description{
  Parse route (with parameters) into regular expression.
}
\value{\item{match}{regular expression to match route}
\item{params}{parameter names}}
\arguments{
  \item{route}{route url.}
}
